+++ bee-block/tests/transaction_essence.rs
@@ -6,9 +6,9 @@ use bee_block::{
     input::{Input, UtxoInput},
     output::{unlock_condition::AddressUnlockCondition, BasicOutput, Output},
     payload::transaction::{RegularTransactionEssence, TransactionEssence, TransactionId},
+    rand::output::rand_inputs_commitment,
     Error,
 };
-use bee_test::rand::output::rand_inputs_commitment;
 use packable::{error::UnpackError, PackableExt};
 
 const TRANSACTION_ID: &str = "0x52fdfc072182654f163f5f0f9a621d729566c74d10037c4d7bbb0407d1e2c649";